[delete] delete todos.ts /w fetch requests, rename axiostodo.ts to todos.ts
